Refactor design and components
