--- src/components/atoms/ArticlesPreviewWrapper.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const ArticlesPreviewWrapper = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  align-items: center;
-  align-content: space-between;
-`;
