+++ src/components/atoms/index.ts
@@ -1,5 +1,4 @@
 export * from './MyLinks';
-export * from './SocialLinks';
 export * from './TagList';
 export * from './Paginator';
 export * from './SEOMeta';
