+++ src/components/organisms/ArticleList.tsx
@@ -1,25 +1,26 @@
 import React from 'react';
-import {ArticlesPreviewWrapper} from '../atoms/ArticlesPreviewWrapper';
-import {ArticlePreviewCard} from '../molecules';
+import styled from 'styled-components';
+import {ArticleCard} from '../molecules';
 import {MarkdownRemarkEdge} from '../../types';
 
 type Props = {
   postEdges: MarkdownRemarkEdge[];
-  description?: React.ReactNode;
 };
 
-export const ArticleList: React.FC<Props> = ({postEdges, description}) => {
+export const ArticleList: React.FC<Props> = ({postEdges}) => {
   return (
-    <div>
-      {description}
-      <ArticlesPreviewWrapper>
-        {postEdges.map((edge) => (
-          <ArticlePreviewCard
-            key={edge.node.frontmatter.title}
-            postInfo={edge}
-          />
-        ))}
-      </ArticlesPreviewWrapper>
-    </div>
+    <Wrapper>
+      {postEdges.map((edge) => (
+        <ArticleCard key={edge.node.frontmatter.title} postInfo={edge} />
+      ))}
+    </Wrapper>
   );
 };
+
+const Wrapper = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  align-items: center;
+  align-content: space-between;
+`;
