+++ src/components/organisms/BaseArticlePageLayout.tsx
@@ -2,14 +2,16 @@ import React from 'react';
 import styled from 'styled-components';
 
 type Props = {
+  articleHeader?: React.ReactNode;
   articles: React.ReactNode;
   profile: React.ReactNode;
   pagination: React.ReactNode;
 };
 
-export const BaseArticlePageLayout: React.FC<Props> = ({articles, profile, pagination}) => {
+export const BaseArticlePageLayout: React.FC<Props> = ({articleHeader, articles, profile, pagination}) => {
   return (
     <Wrapper>
+      {articleHeader && <ArticleHeader>{articleHeader}</ArticleHeader>}
       <ArticleArea>{articles}</ArticleArea>
       <PaginationArea>{pagination}</PaginationArea>
       <ProfileArea>{profile}</ProfileArea>
@@ -25,6 +27,10 @@ const Wrapper = styled.div`
   align-content: space-between;
   flex-wrap: wrap;
   margin: 0 ${(props) => props.theme.layout.marginVertical}px;
+`;
+
+const ArticleHeader = styled.div`
+  width: 100%;
 `;
 
 const ArticleArea = styled.div`
