+++ src/components/templates/CategoryTemplate.tsx
@@ -28,16 +28,12 @@ export const CategoryTemplate: React.FC<Props> = ({pageContext, data}) => {
       </Helmet>
 
       <BaseArticlePageLayout
-        articles={
-          <ArticleList
-            postEdges={postEdges}
-            description={
-              <div style={{margin: '0 30px'}}>
-                <CategoryBanner categoryName={category} />
-              </div>
-            }
-          />
+        articleHeader={
+          <div style={{margin: '0 30px'}}>
+            <CategoryBanner categoryName={category} />
+          </div>
         }
+        articles={<ArticleList postEdges={postEdges} />}
         pagination={<Paginator context={pageContext} />}
         profile={<AuthorCard />}
       />
