+++ src/theme.ts
@@ -1,6 +1,6 @@
 import {colors} from '../data/color';
 
-export const theme= {
+export const theme = {
   color: {
     primaryVivid: colors.primaryVivid,
     white: colors.white,
@@ -10,14 +10,14 @@ export const theme= {
   layout: {
     marginVertical: 20,
     autherAvatorSizeNum: 60,
+    articleCardLogoSize: 60,
   },
   fontFamily: {
-    primary: `-apple-system, BlinkMacSystemFont, Roboto, "Segoe UI", "Helvetica Neue", HelveticaNeue, YuGothic, "Yu Gothic Medium", "Yu Gothic", Verdana, Meiryo, sans-serif`
+    primary: `-apple-system, BlinkMacSystemFont, Roboto, "Segoe UI", "Helvetica Neue", HelveticaNeue, YuGothic, "Yu Gothic Medium", "Yu Gothic", Verdana, Meiryo, sans-serif`,
   },
   fontSize: {
     tmp_1: 24,
     tmp_2: 12,
-    
     tmp_3: 54,
   },
 } as const;
